replace 'NA' placeholder with empty field
In case an expected field is not supplied by the API's 'api_phy' debugfs
file, the placeholder 'NA' was previously used to denote the information
is not available. However, this should not be necessary for parsers
cause they should be able to recognize empty fields as missing
information. Thus, the placeholder was just overhead.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>