Remove 'rc_mode' from expected 'api_phy' input
In the newest API revision, 'rc_mode' is no longer valid for the whole
phy but rather can be set per STA. Thus, this information no longer
belongs to the API's per-phy output.
The current 'rc_mode' is now included in the 'sta;add' and 'sta;dump;
lines issued by the API. The same applies to the new 'tpc_mode'.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>