add new RC mode identifier to phy;add lines
Recent changes in ORCA UAPI allow it to also operate with a subset
of its functionality on hardware that does not use Minstrel as RC. For
such hardware, a lot of information is not available and the
representation of transmission rates in TX status lines has changed.

To make sure clients use the correct format, the UAPI and orca-rcd
supply a new field that determines where RC is handled. This value is
printed in the phy;add lines before the driver and can be either 'fw' or
'mac', meaning that the PHY has its own RC algorithm running in firmware
or relies on the mac80211 RC algorithm (Minstrel) respectively.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>